+++ BCI_analysis/__init__.py
@@ -1,3 +1,3 @@
 from .io import io_matlab, io_python, io_pybpod, io_utils, io_scanimage, io_suite2p
-from .pipeline import pipeline_bpod, pipeline_imaging
+from .pipeline import pipeline_bpod, pipeline_imaging, pipeline_utils
 from .plot import plot_imaging
