+++ BCI_analysis/pipeline/pipeline_imaging.py
@@ -196,72 +196,3 @@ def find_conditioned_neuron_idx(session_bpod_file,session_ops_file,fov_stats_fil
     return cond_s2p_idx,closed_loop_trial,scanimage_filenames
 
 
-def remove_stim_artefacts(F,Fneu,frames_per_file):
-    """
-    removing stimulation artefacts with linear interpolation
-    and nan-ing out tripped PMT traces
-    Assumes that files start with photostim.
-
-    Parameters
-    ----------
-    F : matrix of float
-        Fluorescence of ROIs
-    Fneu : matrix of float
-        fluorescence of neuropil
-    frames_per_file : list of int
-        # of frames in each file (where the photostim happens)
-
-    Returns
-    -------
-    F : matrix of float
-        corrected fluorescence of ROIs
-    Fneu : matrix of float
-        corrected fluorescence of neuropil
-
-    """
-    artefact_indices = []
-    fneu_mean = np.mean(Fneu,0)
-    
-    for stim_idx in np.concatenate([[0],np.cumsum(frames_per_file)[:-1]]):
-        idx_now = []
-        if stim_idx>0 and fneu_mean[stim_idx-2]*1.1<fneu_mean[stim_idx-1]:
-            idx_now.append(stim_idx-1)
-        idx_now.append(stim_idx)
-        if stim_idx<len(fneu_mean)-2 and fneu_mean[stim_idx+2]*1.1<fneu_mean[stim_idx+1]:
-            idx_now.append(stim_idx+1)
-        artefact_indices.append(idx_now)
-    
-    f_std = np.std(F,0)
-    pmt_off_indices = f_std<np.median(f_std)-3*np.std(f_std)
-    pmt_off_edges = np.diff(np.concatenate([pmt_off_indices,[0]]))
-    pmt_off_indices[pmt_off_edges!=0] = 1 #dilate 1
-    pmt_off_edges = np.diff(np.concatenate([[0],pmt_off_indices,[0]]))
-    starts = np.where(pmt_off_edges==1)[0]
-    ends = np.where(pmt_off_edges==-1)[0]
-    lengths = ends-starts
-    for idx in np.where(lengths<=10)[0]:
-        pmt_off_indices[starts[idx]:ends[idx]]=0
-    pmt_off_edges = np.diff(np.concatenate([[0],pmt_off_indices,[0]]))
-    starts = np.where(pmt_off_edges==1)[0]
-    ends = np.where(pmt_off_edges==-1)[0]   
-    for end_ in ends:
-        pmt_off_indices[end_:end_+10] = 1 # 10 extra frames are added at the end of pmt trips to be removed
-        
-    
-    
-    F_ = F.copy()
-    F_[:,np.concatenate(artefact_indices)]=np.nan
-    for f in F_:
-        nans, x= nan_helper(f)
-        f[nans]= np.interp(x(nans), x(~nans), f[~nans])
-        f[pmt_off_indices] = np.nan
-    F = F_
-    
-    Fneu_ = Fneu.copy()
-    Fneu_[:,np.concatenate(artefact_indices)]=np.nan
-    for f in Fneu_:
-        nans, x= nan_helper(f)
-        f[nans]= np.interp(x(nans), x(~nans), f[~nans])
-        f[pmt_off_indices] = np.nan
-    Fneu = Fneu_
-    return F, Fneu
