+++ BCI_analysis/pipeline/pipeline_utils.py
@@ -1,5 +1,5 @@
 import numpy as np
-
+import warnings
 def nan_helper(y):
     """Helper to handle indices and logical indices of NaNs.
 
@@ -28,23 +28,29 @@ def rollingfun(y, window = 10, func = 'mean'):
     @input:
         y = array, window, function (mean,min,max,std)
     """
-    window = int(window)
-    if window >len(y):
-        window = len(y)-1
-    y = np.concatenate([y[window::-1],y,y[:-1*window:-1]])
-    ys = list()
-    for idx in range(window):    
-        ys.append(np.roll(y,idx-round(window/2)))
-    if func =='mean':
-        out = np.nanmean(ys,0)[window:-window]
-    elif func == 'min':
-        out = np.nanmin(ys,0)[window:-window]
-    elif func == 'max':
-        out = np.nanmax(ys,0)[window:-window]
-    elif func == 'std':
-        out = np.nanstd(ys,0)[window:-window]
-    elif func == 'median':
-        out = np.nanmedian(ys,0)[window:-window]
-    else:
-        print('undefinied funcion in rollinfun')
+    
+
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", category=RuntimeWarning)
+    
+        window = int(window)
+        if window >len(y):
+            window = len(y)-1
+        y = np.concatenate([y[window::-1],y,y[:-1*window:-1]])
+        ys = list()
+        for idx in range(window):    
+            ys.append(np.roll(y,idx-round(window/2)))
+        if func =='mean':
+            out = np.nanmean(ys,0)[window:-window]
+        elif func == 'min':
+            out = np.nanmin(ys,0)[window:-window]
+        elif func == 'max':
+            out = np.nanmax(ys,0)[window:-window]
+        elif func == 'std':
+            out = np.nanstd(ys,0)[window:-window]
+        elif func == 'median':
+            out = np.nanmedian(ys,0)[window:-window]
+        else:
+            print('undefinied funcion in rollinfun')
     return out
